Power-Activation-Gateway/requirements.txt
PyYAML
Flask
pytz
celery
Flask-SocketIO